+++ frontend/src/components/ResourceDisplay.vue
@@ -10,7 +10,7 @@
             <v-col cols=12 v-if="!collapsed[key] && ( (resource.schema && resource.schema.fields) || (resource.tableSchema && resource.tableSchema.fields))">
                 <div v-for="(field, fkey) in ((resource.tableSchema && resource.tableSchema.fields) ? resource.tableSchema.fields : resource.schema.fields)" :key="'field-'+fkey+'-'+(field ? field.name : '')" :class="`field` + displayClass('resources.' + key + '.schema.fields.' + fkey.toString())">
                     <v-row v-if="field && field.name" :class="'ma-0 noOverflow' + displayClass('resources.' + key + '.schema.fields.' + fkey.toString() + '.name')">
-                        <v-col cols=1>
+                        <v-col cols=1 v-if="collapsedFields && collapsedFields[key] && collapsedFields[key][fkey]">
                             <v-btn @click="collapse(key, fkey)"><v-icon>{{collapsedFields[key][fkey] ? 'mdi-plus' : 'mdi-minus'}}</v-icon></v-btn>
                         </v-col>
                         <v-col cols=2></v-col>
@@ -19,7 +19,7 @@
                         </v-col>
                     </v-row>
 
-                    <span v-if="!collapsedFields[key][fkey]">
+                    <span v-if="!collapsedFields || !collapsedFields[key] || !collapsedFields[key][fkey]">
                         <v-row v-for="(attribute, attribKey) in sortedField(field, key, fkey)" :key="'field-'+fkey+'-'+attribKey+'attribute'" :class="'ma-0' + displayClass('resources.' + key + '.schema.fields.' + fkey.toString() + '.'+attribKey)">
                             <v-col cols="3" v-if="attribKey !== 'name' && attribKey !== 'highlight'">
                                 {{$tc(attribKey)}}
@@ -106,24 +106,28 @@ export default {
     methods: {
 
         sortedField: function(field, key, fkey){
+            let f = JSON.parse(JSON.stringify(field));
             if (this.other && this.other[key] && this.other[key].schema && this.other[key].schema.fields && this.other[key].schema.fields[fkey]){
                 let otherAttribs = Object.keys(this.other[key].schema.fields[fkey]);
                 for (let i=0; i<otherAttribs.length; i++){
-                    if (typeof(field[otherAttribs[i]]) === 'undefined'){
-                        field[otherAttribs[i]] = "<Not Specified>";
+                    
+                    if (f && typeof(f[otherAttribs[i]]) === 'undefined'){
+                        f[otherAttribs[i]] = "<Not Specified>";
                     }
                 }
             }
 
-            field = Object.keys(field).sort().reduce(
-                (obj, key) => { 
-                    obj[key] = field[key]; 
-                    return obj;
-                }, 
-                {}
-            );
+            if (f){
+                f = Object.keys(f).sort().reduce(
+                    (obj, key) => { 
+                        obj[key] = f[key]; 
+                        return obj;
+                    }, 
+                    {}
+                );
+            }
 
-            return field;
+            return f;
 
         },